BIIN-601 Add new look to element mini view header
BIIN-601
Add new look to element mini view header

+++ biin/biin/models/BNElement.swift
@@ -52,7 +52,7 @@ class BNElement:NSObject {
     var reservedQuantity:String?
     var claimedQuantity:String?
     var actualQuantity:String?
-    var stars:Int = 0
+    var stars:Float = 0
     var useWhiteText = false
     
     //Details
