+++ biin/biin/models/BNSite.swift
@@ -23,7 +23,7 @@ class BNSite:NSObject, NSCoding {
     
     //Color
     var titleColor:UIColor?
-    var stars:Int = 0
+    var stars:Float = 0
     
     //Location
     var country:String?
